move sb utils to src
