+++ packages/slice-machine/lib/utils/str.ts
@@ -32,28 +32,3 @@ export function createStorybookId(str: string): string {
   const camel = camelCase(str);
   return `_${camel[0].toUpperCase()}${camel.slice(1)}`;
 }
-
-export function camelCaseToDash(v: any): string {
-  let ret = "",
-    prevLowercase = false,
-    prevIsNumber = false,
-    isFirstChar = true;
-
-  for (const s of v) {
-    const isUppercase = s.toUpperCase() === s;
-    const isNumber = !isNaN(s);
-
-    if (isNumber && !prevIsNumber && !isFirstChar) {
-      ret += "-";
-    } else {
-      if (isUppercase && !isNumber && (prevLowercase || prevIsNumber)) {
-        ret += "-";
-      }
-    }
-    ret += s;
-    prevLowercase = !isUppercase;
-    prevIsNumber = isNumber;
-    isFirstChar = false;
-  }
-  return ret.replace(/-+/g, "-").toLowerCase();
-}
