+++ packages/slice-machine/src/utils/storybook.ts
@@ -1,4 +1,29 @@
-import { createStorybookId, camelCaseToDash } from "./str";
+import { createStorybookId } from "@lib/utils/str";
+
+function camelCaseToDash(v: any): string {
+  let ret = "",
+    prevLowercase = false,
+    prevIsNumber = false,
+    isFirstChar = true;
+
+  for (const s of v) {
+    const isUppercase = s.toUpperCase() === s;
+    const isNumber = !isNaN(s);
+
+    if (isNumber && !prevIsNumber && !isFirstChar) {
+      ret += "-";
+    } else {
+      if (isUppercase && !isNumber && (prevLowercase || prevIsNumber)) {
+        ret += "-";
+      }
+    }
+    ret += s;
+    prevLowercase = !isUppercase;
+    prevIsNumber = isNumber;
+    isFirstChar = false;
+  }
+  return ret.replace(/-+/g, "-").toLowerCase();
+}
 
 export const sanitizeSbId = (str: string) => {
   return str
